requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Optional: Local embeddings for semantic code search
# Only needed if LOCAL_EMBEDDINGS_ENABLED=true
sentence-transformers>=2.2.0
chromadb>=0.4.0